+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Set the default `cloudChannel` to `8`. #INT-3350
+
 ## 6.2.1 - 2025-06-03
 
 ### Fixed